Moved collaboration to super_archives and plugins
