+++ colab/home/views.py
@@ -2,7 +2,7 @@ from django.conf import settings
 from django.shortcuts import render
 from django.http import HttpResponse, Http404
 
-from colab.search.utils import get_collaboration_data
+from colab.plugins.utils.collaborations import get_collaboration_data
 from colab.super_archives.models import Thread
 from colab.accounts.utils import mailman
 from colab.accounts.models import User
